Added swap stack feature. Currently uses mod to switch between one of two options, however we can easily addapt to use the label of a given button as string input for the IndexOf() method.
Also, see lines 70-72 in MainWindow.py to see where I have hard-coded the names of what our objects are called. We can put these in a constants.py file or an object somewhere, and this is our hashmap of our keys to index with. i.e. IndexOf(object.label: str) == 1, then set the stacks CurrentIndex to 1.